+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # 9.0.0
 
-* Migrate to 2.10.0
+* Migrate to 2.10.0.
+* Add shouldShowSelectionHandles and textSelectionGestureDetectorBuilder call back to define the behavior of handles and toolbar.
+* Shortcut support for web and desktop.
 
 # 8.0.2